Remove usage of dynamicType
Using dynamicType to get the bundle for a class (obviously) returns the bundle of the class that the original class was replaced with. This caused an error where icons would not be displayed when a user replaced a class with his own. All instances of dynamicType have now been replaced with the actual class.